NHIT-Backend-API-Tests/python-pytest/requirements.txt
pytest==7.4.3
pytest-html==4.1.1
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-cov==4.1.0
requests==2.31.0
python-dotenv==1.0.0
faker==20.1.0
pydantic>=2.5.2
allure-pytest==2.13.2
